Remove compiler warnings
Fix all compilation warnings (-Wall -Wextra):
  - fem_2d.hpp: Remove unused variables 'a' and 'b' in local_stiffness_q1()
  - fem_2d.hpp: Cast n_dof to size_t in vector size comparison
  - matrix.hpp: Cast cols_ to size_t in vector size comparison
  - solver.hpp: Cast n to size_t in vector size comparison

Build now completes cleanly without warnings (except C++17 extension notes).

+++ src/matrix.hpp
@@ -39,7 +39,7 @@ public:
 
     // Multiplication matrice-vecteur: y = A*x
     std::vector<double> multiply(const std::vector<double>& x) const {
-        if (x.size() != cols_) {
+        if (x.size() != static_cast<size_t>(cols_)) {
             throw std::invalid_argument("Vector size mismatch");
         }
         std::vector<double> y(rows_, 0.0);
